Fix precedence bugs.
This commit fixes two issues:

1. There was a bug where the Nativity of St. John was taking precedence
over the Feast of the Sacred Heart and Corpus Christi.  (The former case
occurred in 2022.)  Now The Nativity of St. John will be transferred to
June 23 if such a conflict exists.

2. The vigils of second class feasts were not being handled correctly.
The vigil of a second class feast would occasionally outrank another
second class feast.  This bug has also been fixed.

In addition to this I added a helper function to iterate over the
liturgical year and made some minor stylistic changes in a few places.